ipxe: move fuzzer to separate file
